tools/v2check: reorg code
