+++ tools/v2check/v2check/v2check.go
@@ -1,9 +1,9 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016 Datadog, Inc.
+// Copyright 2023 Datadog, Inc.
 
-package main
+package v2check
 
 import (
 	"errors"
@@ -14,7 +14,9 @@ import (
 	"golang.org/x/tools/go/ast/inspector"
 )
 
-var analyzer = &analysis.Analyzer{
+var knownChanges = []*knownChange{}
+
+var Analyzer = &analysis.Analyzer{
 	Name:     "v2check",
 	Doc:      "Migration tool to assist with the dd-trace-go v2 upgrade",
 	Requires: []*analysis.Analyzer{inspect.Analyzer},
